Rename react to react-components
