+++ core/utils/poly-look/src/react-components/legends/blockLegend/blockLegend.jsx
